Fix phone call state not always updating
TelephonyManager.listen() only uses a field of the PhoneStateListener,
and that field only has weak references to the PhoneStateListener itself.
Our PhoneStateListener was being garbage-collected. We need to keep
a reference to it to prevent that.

This fixes issues like the call state not changing from "ringing" to
the call duration, and the information not going away when the call is
over.

Bug: None
Test: Manually
Change-Id: Ib26afe61498a50ff7b834c3bc4adea5ddd939905