Make cluster stop spamming accessibility events
The cluster was causing CTS tests to fail, because
they use waitForIdle(). waitForIdle() waits for a lack
of accessibility events to be sent out, and fails the
test if there isn't ever a quiet period.

Cluster's sensor readings were constantly being updated,
causing lots of accessibility events. Check that
the TextView actually needs to be updated before
updating it.

Bug: 182417178
Test: Manually
Change-Id: I51f4e60c4caaebf7601b098011fb16bfc46db175