Mark FakeFreeNavigationActivity showForAllUsers.
When DirectRenderingCluster is selected for the system cluster role,
it can give the permissions for the system package only.
So we should install DirectRenderingCluster only for the system user,
or it fails DefaultAutoPermissionGrantPolicyTest.

DirectRenderingCluster has two Activities, MainClusterActivity and
FakeFreeNavigationActivity.
MainClusterActivity is running as user0, but FakeFreeNavigationActivity
is supposed to run as the current user.

To make FakeFreeNavigationActivity work in the system user, this CL
declares the Activity showForAllUsers.
In addition, to make ClusterActivity accommodate NavigationActivity
for both user0 and user10, this checks the resolved ActivityInfo.flag to
determine if the Activity is for user0 or user10.

Bug: 190862565
Test: check if DirectRenderingCluster works correctly with the different configs of freeNavigationIntent and navigationOnly.
Change-Id: Ia9501ba67f8e3915ce9037764ff957e724d05a01